Models Work, and then some cleanup (#12)
* Match func_8003B640

* Fix mmAlloc2 prototype

* Match modInitModels

* Add some names

* MAtch func_8003BED0

* Fix some vscode stuff to make the syntax highlighting better, including use clangd if it's installed.

* Fixes so generating the ctx.c file will work.

* Fix for all warnings in ctx.c found with duplicate definitions

* Formatter and more warning fixes

* Near match modFreeModel

* Match...modFreeModel

* Assign correct blame

* Kind of match texFreeTexture, but the struct needs to be reworked.

* Small changes

* Match mainDemoOnly

* Sad, but I can no longer blame bloodangel :(

* Do a good effort on RevealReturnAddresses

* MAtch func_8003C6D0

* formatter

* Move some structs

+++ src/diprint.c
@@ -1,5 +1,6 @@
 #include "diprint.h"
 #include "common.h"
+#include "gameVi.h"
 #include "libc/string.h"
 #include "stdarg.h"
 
@@ -970,7 +971,7 @@ void diPrintfAll(Gfx **dList) {
     gDebugScreenHeight = height;
     gDPSetScissor((*dList)++, 0, 0, 0, gDebugScreenWidth, gDebugScreenHeight);
     debug_text_bounds();
-    gSPDisplayList((*dList)++, dDebugFontSettings);
+    gSPDisplayList((*dList)++, (u32) dDebugFontSettings);
     buffer = (char *) gDebugPrintBufferStart;
     debug_text_origin();
     gDebugFontTexture = -1;
